scancode avec possibilité d'effacer

+++ src/math.cpp
@@ -1,6 +1,6 @@
 #include "math.hpp"
 
-int power_int(int x, int y) {
+uint64_t power_int(uint32_t x, uint32_t y) {
     int res = 1;
     while (y-- > 0)
         res *= x;
