+++ src/math.hpp
@@ -1,6 +1,8 @@
 #ifndef __MATH_H_
 #define __MATH_H_
 
-int power_int(int x, int y); //compute x^y
+#include "types.hpp"
+
+uint64_t power_int(uint32_t x, uint32_t y); //compute x^y
 
 #endif
